Added detailed messages to settings cloning
Added ability to eject unpackager trackers
Added ability to clear recipe holders and markers in crafting grid
Changed saving settings to cloner to use shift-right click

+++ src/main/java/thelm/packagedauto/client/screen/ItemAmountSpecifyingScreen.java
@@ -44,7 +44,7 @@ public class ItemAmountSpecifyingScreen extends AmountSpecifyingScreen<ItemAmoun
 			int amount = MathHelper.clamp(Integer.parseInt(amountField.getValue()), 0, maxAmount);
 			ItemStack newStack = stack.copy();
 			newStack.setCount(amount);
-			PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket((short)containerSlot, newStack));
+			PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket(containerSlot, newStack));
 			close();
 		}
 		catch(NumberFormatException e) {
